refactor(mapping): name the deletion module after its purpose, clarify a comment
`existing_mappings.py` described only the module's first half and read like a data
holder. The module answers one question for `mapping_add`: which of the ontology's
existing external mappings must be deleted. It both reads the current mappings out
of the ontology turtle and diffs them against the Excel file, so it is now
`find_mappings_to_delete.py`, with the test module renamed to match. The functions,
constants and `ExistingMappings` keep their names: they describe the data that was
read, which is still accurate.

The comment guarding `NO_MAPPING_IN_INPUT` used "mapping" in two senses in one
sentence - the Excel input and the state on the server - which made it read as if an
entity with no mappings could lose its mappings. It now names both sides explicitly.
