+++ src/dsp_tools/commands/mapping/resolve_parsed_mappings.py
@@ -47,8 +47,10 @@ def _resolve_one_mapping(
     resolved: list[str] = []
     problems: list[PrefixResolutionProblem] = []
     if not parsed_mapping.prefixed_mapping_iris:
-        # An entity without any mapping would be stripped of its existing mappings and get nothing in return,
-        # so this must fail before the upload phase.
+        # The mapping cell of this Excel row resolves to no IRI at all, e.g. a cell containing only ";".
+        # That is a hazard under replace semantics: the delete phase would strip every external mapping the
+        # entity currently has on the server, while the add phase could only send an empty list, which the
+        # DSP-API rejects. The entity would end up with no external mappings at all, so fail before any write.
         problems.append(
             PrefixResolutionProblem(
                 entity_name=parsed_mapping.name,
